Sign warnings

+++ src/DirectSolver/DirectSolver-COO-MUMPS-Give/applySymmetryShift.cpp
@@ -123,7 +123,7 @@ void DirectSolverGive::applySymmetryShiftOuterBoundary(Vector<double> x) const
 // clang-format off
 void DirectSolverGive::applySymmetryShift(Vector<double> x) const
 {
-    assert(x.size() == grid_.numberOfNodes());
+    assert(x.size() == static_cast<uint>(grid_.numberOfNodes()));
     assert(grid_.nr() >= 4);
 
     if (num_omp_threads_ == 1) {
